Modified main.py
Updated the method spawnTrash and reorganized the main loop in a way that I can change the level designs more easily. Now I can change the background and trash spawn rate depending on the currentLevel. For example, at level 2 the background is completely white and only one trash spawns at a time.In addition, the timer does not run. In the future, I plan to create another list which are trash collect goals for the player to follow. If the player meets the goal, less trash will spawn on the map and the map visuals will change. Else the opposite effect will occur.