Level Creation Enhancements
Updating the main loop in main.py to change the trash spawn rate base on the player's performance in the level. Using a dictionary called levelsList, it holds both the levels and the min collection goal for each level. As a result, I am able to create a mini tutorial from this feature and also change the level design all from the player's performance. In addition, I started prepping for a 'Level Complete ' screen after the level is done. Currently there are no visuals other than a purple screen and it cannot be accessed through normal gameplay. I will continue to tweak this performance base feature, then implement the new state