Level Selection Beta
Levels are now able to end at certain conditions rather than run infinitely. In addition, when selecting the Prologue level, the player will play an optional tutorial. After that, they are able to play level one. After the level is finished, the score will show at the end and the next level would play. At the moment, there is only one level, so after the level is finished, the player will be sent back to the main menu. They can replay the level again and override their old scores. At the moment, all upgrades purchase will stay with the player no matter what level they finish. As a result, I plan to change the cost of the upgrades in the future. My next step at the moment is to clean up the code and add comments to the new additions made.