Added New Treasure Sprite
Created and added the treasure sprite in the game. Due to the size, the sprite is very pixelated. I must enlarge the sprite further to display properly. For a moment, I encountered a bug where the player moves very fast despite no upgrades and changes to the actual speed. At first, I thought it was because of a misinput of the speed somewhere in the main.py or level.py files. After checking with no success, I thought about the clock speed and cache. However, I was able to fix this after uncommenting some code in the beginning of the main loop (delay 100). This seemed to keep the framerate/clock speed of the display in check.