Added Tutorial Background
Created and added a background for the tutorial. After adding the image in window.py, I decided to modify all the tuples to hold the boundaries. These boundaries are used to block the player from walking in certain areas. With this update, I switched the constants in the main.py with the backgroundList items from window.py