BugFix: Treasure Sprite Lag
When the treasure sprite spawns and the player tries to collect it, the game lags for a second. My initial thought is that the image file was too large due to the details, causing performance errors. To test this, I switched the sprite with a simpler sprite and saw that there was a significance improvement when the other sprite was used. With a quick search, I found that the main error was that the way I loaded the images took too much time and storage. Everytime a new trash object is initialized, the trash sprites ALWAYS reloaded, which took a bunch of resources. The issue was more obvious with the treasure sprite due to its size. To fix this, I created a method in the trash.py to loop through the list, and replace each entry with a loaded version of the image. I then call this method once in main.py. I still have a lot to learn on optimization