[Ongoing] Fail Screen
I cleaned up the score feature and started to implement a fail screen when the player does not collect the minimum amount of trash in the whole level. This minimum is the average of all scores in the levelsList. At the moment, if the final score is less than the average, then a pink screen in shown. Else the default end screen is shown. The next step is to properly create the ending screen in Canva and import it in the game.