Developing Level/Scene Two
Started to implement the foundations for level two. I created a sceneTwo class in level.py to start developing the next cutscene. For now, I simply displayed the player
and use SPACE to finish the cutscene as a placeholder. I then tried to implement the levelTwo foundations in main.py main loop. This is done by adding a list to the levelsList and the scoresList. This allows the level to be played properly. I then worked around the code in both level.py and main.py to get the order of the LevelSelection screen to work properly and display the correct endings (ln 277-308 in main.py & gameEnd class in level.py)