[Ongoing] Level Two Foundations Updates
I tried to play through the game with the new additions I added. However, when playing from the beginning (Prologue), the player finishes the tutorial but does not go to the first level. After an hour moving the code around and seeing if the variables are updated properly, I found the main issue was the failScore calculation. By default, I made it totalScore/3 instead of totalScore / totalRounds. As a result, when the player finishes the tutorial, they are unable to meet the pass condition (5/3) since I made it CLEAR the totalScore after the tutorial is complete. To work around this, I simply allow the player to keep the 5 points scored from the tutorial. At the final screen, I will simply do a for loop that skips the first index when calculating the score, or keep it as a bonus for players who did the tutorial.